rename LS message
